Chore: add heroku postgres
-install make runnable
-configure load of database to heroku

[Finishes #165405232]
